+++ CHANGELOG.md
@@ -4,7 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Fix reponsive font-size for `CounterBlock` component.
+- Fix: Update `CrowdfundingCard` props.
+
+## [23.7.0] - 2018-10-24
+
+Feature:
+- Add `itemHeight` props to `Timeline` component.
+
+Fix:
+- Fix reponsive font-size for `CounterBlock` component.
 
 ## [23.6.0] - 2018-10-19
 
@@ -42,6 +50,9 @@ Fixes:
 - Update `ffi` gem.
 - Deprecate `ContributionCard` component. You should use `RewardCard`
   component instead. It will be deleted in further versions.
+- Feature: Add `ExpandBoard`, `ExpandBoard.Button` and `ExpandBoard.Content`
+  components.
+- Feature: Add `List` and `List.ButtonItem` components.
 
 ## [23.5.1] - 2018-09-27